Rewrite all the things
First commit of major rewrite. Goal is to fix the awful error handling,
introduce more tests, simplify the code and just clean up this mess.

A full changelog will come with the final release but here are some of
the points that have been reworked so far:
 * Add tests
 * Add command line arg parser (clap)
 * Rewrite main
 * Rewrite utilities and templates
 * Combine Sqlite and Templates
 * Introduce proper error handling

Things that still need to be worked on:
 * Improve path resolving for stuff like "$HOME"
 * Make every test green again!
 * SQLite templater
 * Loads I've probably forgotten