Add SQLite templating
The SQLite templating has been reworked and added.
Currently it's using Rust's `format!` instead of `?` for constructing
the SQLite queries which works but hopefully can be changed in the
future.
The last part missing from this new Implementation is the
SQLite-Templater for creating a templated SQLite DB from any SQLite DB.
Everything has been automatically and manually tested but the Test
dataset is small and doesn't cover a lot of scenarios so it might be
necessary to test more edge cases and error behavior.