Add Template Creator Feature
Added the ability to generate the templated files and databases
automatically so editing can be done with the real files.
By default this is first done using the "templates/old.toml" file which
is created after running the dotfiler and if that doesn't exist it uses
the default "config.toml" instead.
The SQLite Template Creator has been added but there hasn't been any
automatic or manual testing yet.
Most of the stuff that still has to be done is annotated as TODO in main
and template files. Generally better test coverage might be a good idea.