Add "add" command
Added a new command that allows the user to automatically add a file,
directory or symlink to the dotfiler. A bunch of stuff still needs a lot
of work which is documented in several "TODO"s.
Adding directories is manually and automatically tested, files only
manually and symlinks not at all yet. Once this feature exists cleanly
on error and the config doesn't get blown up by adding the same thing
multiple times the first usable version might get released.