Add backups/cache to Symlinks
For creating a safe dotfile solution the best option is to always store
the last version of the dotfiles in a cache/backup directory. That
allows the user to manually roll back changes. These also can be synced
with git automatically for version control.
This should also be used by the recursive `restore()` method so files
are restored from the cache/backup not from data stored in memory, which
reduces the amount of memory required.

This change has been implemented and tested for Symlinks only so far.
Directories are excluded from caching, so now Files are left. This also
allows the reimplementation of the SQLite database in the near future.