Add BinaryFile and TextFile caching
A new BinaryFile has been added to filesystem.rs which is not rendered
nor templated for files which are just copied without any modification.
The method for detecting binary files tries to read a file as utf8 and
whenever that fails it will declare the file as binary.

The TextFile has been modified to save to cache so the user is able to
restore and keep track of the last revision manually.
The restore method itself also now relies on the cached version of the
old file.