Add ability to execute scripts post templating
It is now possible to execute scripts after templating. This can execute
any script that is executable using "sh". This was tested with both bash
and python scripts and worked successfully.
Output and Stderr are redirected to console to allow the user to check
for any errors or issues, if one script fails the others are still
executed.